+++ NEWS.md
@@ -3,6 +3,14 @@ This file documents the development of the package as well as open issues or poi
 
 
 
+### Version in 1.1.7
+Date: 2019.02.17 (current development)
+
+Changes:
+*   bestModel function noew accepts a formula as a predictors object
+
+
+
 ### Version in 1.1.6
 Date: 2019.02.07, third release on CRAN